sm8150-common: add wifi mac generator
The default qcacld-3.0 implementation looks for the MAC
in a firmware with the name `wlan/qca_cld/wlan_mac.bin`.
That firmware is, by default, symlinked to
`/mnt/vendor/persist/wlan_mac.bin`.
Xiaomi has somewhat changed the qcacld-3.0 logic to load
the MAC from another place. Xiaomi's nvmac puts a plain-text
MAC in `/data/vendor/mac_addr/wlan_mac.bin`.
Add a script to write the MAC address found in the file
created by nvmac to the file in persist that the firmware
file is symlinked to.

This implementation is (loosely) based on the one
written for OnePlus 7.

Change-Id: I8e22ddbc8a8b6b5f493a32ab179e48c1815e19bc